+++ src/main/java/io/github/astrapi69/crypt/data/certificate/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package provides classes that holds information for certificates
- */
-package io.github.astrapi69.crypt.data.certificate;